+++ docker/nginx.conf
@@ -14,13 +14,13 @@ http {
 
     location / {
       proxy_pass http://frontend-container:82;
-        proxy_set_header Host $http_host;
+        proxy_set_header      Host            $http_host;
     }
 
     location /api {
       proxy_pass http://backend-container:81;
-        proxy_set_header Host $http_host;
-    }
+        proxy_set_header      Host            $http_host;
+        proxy_set_header      Connection      "upgrade";
   }
